Move processor-specific to wpcom-checkout
