+++ packages/wpcom-checkout/src/index.ts
@@ -37,3 +37,5 @@ export * from './removed-from-cart-item';
 export * from './translate-payment-method-names';
 export * from './is-payment-method-enabled';
 export * from './filter-appropriate-payment-methods';
+export * from './payment-processor-countries-fields';
+export * from './processor-specific';
